+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Current
+
+- Fix: `swap_with_slippage_protection()` interpreted `max_slippage` as percent,
+  even though the doc string says BPS
+- API change: `swap_with_slippage_protection(max_slippage=15)` - change the default Uniswap v3
+  trade slippage tolerance from (unrealistic) 0.1 BPS to 15 BPS.
+
 # 0.22.30
 
 - API change: Handle `wait_and_broadcast_multiple_nodes()` so that it will attempt